backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
langchain==0.0.335
pypdf2==3.0.1
python-multipart==0.0.6
pytesseract==0.3.10
sqlalchemy==2.0.23
numpy==1.26.1
scikit-learn==1.3.2
huggingface_hub==0.20.3
mistralai==1.5.2
python-dotenv==1.0.0

# Simplified requirements removing Rust dependencies
# We'll use simpler NLP tools instead
nltk==3.8.1 